Add shared .env loader utility
